requirements.txt
discord.py
mysqlclient
requests